adjust snapshot, longer timeouts. split in even more files
