+++ test/init.vanilla.e2e.spec.ts
@@ -17,11 +17,11 @@ jest.retryTimes(3);
 describe('Init', () => {
   beforeAll(async () => {
     return deleteAsync(testOutputDir, 6);
-  }, 80000);
+  }, 120000);
 
   afterAll(async () => {
-    await deleteAsync(testOutputDir, 6);
-  }, 80000);
+    return deleteAsync(testOutputDir, 6);
+  }, 120000);
 
   test('[E2E] should create a new project with the necessary files', async () => {
     const projectDir = getCleanProjectDir(testOutputDir, 'full-init');
@@ -43,20 +43,5 @@ describe('Init', () => {
 
     await execa('npm', ['run', 'ci-integration-test'], { cwd: projectDir, stdio: 'inherit' });
     expect(fs.readdirSync(path.resolve(reportsPath, 'backend-integration')).length).toBeGreaterThan(1);
-    expect(fs.readdirSync(path.resolve(reportsPath, 'coverage-reports', 'backend-integration')).length).toBeGreaterThan(1);
-  }, 240000);
-
-  test('[E2E] should create a new project with the necessary files when adding cds', async () => {
-    const projectDir = getCleanProjectDir(testOutputDir, 'full-init-cds');
-    await Init.run([projectDir, '--projectName=testingApp', '--buildScaffold', '--no-analytics', '--addCds']);
-
-    ['.cdsrc.json', 'srv/cat-service.cds', 'db/data-model.cds', 'src/catalogue/catalogue.module.ts']
-      .map(file => path.resolve(projectDir, file))
-      .forEach(path => {
-        expect(fs.existsSync(path)).toBe(true);
-      });
-
-    await execa('npm', ['run', 'cds-deploy'], { cwd: projectDir, stdio: 'inherit' });
-    expect(fs.existsSync(path.resolve(projectDir, 'testingApp.db'))).toBeTrue();
   }, 240000);
 });
