+++ test/utils/scaffold.spec.ts
@@ -17,7 +17,7 @@ jest.retryTimes(3);
 
 import * as fs from 'fs';
 import * as path from 'path';
-import { buildScaffold, shouldBuildScaffold } from '../../src/utils';
+import { shouldBuildScaffold } from '../../src/utils';
 import { deleteAsync, getCleanProjectDir, getTestOutputDir } from '../test-utils';
 
 const testOutputDir = getTestOutputDir(__filename);
@@ -25,11 +25,11 @@ const testOutputDir = getTestOutputDir(__filename);
 describe('Scaffold Utils', () => {
   beforeAll(async () => {
     return deleteAsync(testOutputDir, 6);
-  }, 80000);
+  }, 120000);
 
   afterAll(async () => {
     return deleteAsync(testOutputDir, 6);
-  }, 80000);
+  }, 120000);
 
   it('should determine if scaffold is needed', async () => {
     const projectDir = getCleanProjectDir(testOutputDir, 'should-build-scaffold');
@@ -41,12 +41,4 @@ describe('Scaffold Utils', () => {
     fs.copyFileSync(path.resolve('test', 'nest', 'package.json'), packageJsonPath);
     expect(await shouldBuildScaffold(projectDir, false)).toBe(false);
   });
-
-  test('[E2E] should build the scaffold', async () => {
-    const projectDir = getCleanProjectDir(testOutputDir, 'build-scaffold');
-
-    await buildScaffold(projectDir, false, false);
-
-    expect(fs.readdirSync(projectDir).sort()).toMatchSnapshot();
-  }, 80000);
 });
